requirements.txt
paho-mqtt
pyais
streamz
environs
pydantic
datamodel-code-generator